refactor: rearranging radio contents
