+++ radio/radio.go
@@ -1,7 +1,6 @@
 package radio
 
 import (
-	"fmt"
 	"strings"
 
 	"github.com/charmbracelet/bubbles/key"
@@ -9,96 +8,13 @@ import (
 	gloss "github.com/charmbracelet/lipgloss"
 )
 
+// RadioButtonMsg enables communication down into radio buttons
 type RadioButtonMsg struct {
+	// State is whether or not the radio button is active
 	State bool
-}
 
-type ButtonStyles struct {
-	LeftIndicatorCharacter        string
-	RightIndicatorCharacter       string
-	ActiveLeftIndicatorCharacter  string
-	ActiveRightIndicatorCharacter string
-
-	LeftIndicator        gloss.Style
-	RightIndicator       gloss.Style
-	ActiveLeftIndicator  gloss.Style
-	ActiveRightIndicator gloss.Style
-
-	Label       gloss.Style
-	ActiveLabel gloss.Style
-}
-
-func DefaultButtonStyles() ButtonStyles {
-	return ButtonStyles{
-		LeftIndicatorCharacter:       "○",
-		ActiveLeftIndicatorCharacter: "●",
-
-		LeftIndicator:       gloss.NewStyle().Foreground(DefaultUnfocusedColor),
-		ActiveLeftIndicator: gloss.NewStyle().Foreground(DefaultActiveTabIndicatorColor),
-	}
-}
-
-type Button struct {
-	label string
-
-	active bool
-
-	styles ButtonStyles
-}
-
-func NewButton(label string) Button {
-	return Button{
-		label:  label,
-		styles: DefaultButtonStyles(),
-	}
-}
-
-func (m Button) Styles(styles ButtonStyles) Button {
-	m.styles = styles
-	return m
-}
-
-func (m Button) Init() tea.Cmd {
-	return nil
-}
-
-func (m Button) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
-	switch msg := msg.(type) {
-	case RadioButtonMsg:
-		m.active = msg.State
-	}
-	return m, nil
-}
-
-func (m Button) View() string {
-	lic := m.styles.LeftIndicatorCharacter
-	ric := m.styles.RightIndicatorCharacter
-	li := m.styles.LeftIndicator
-	ri := m.styles.RightIndicator
-	l := m.styles.Label
-
-	if m.active {
-		lic = m.styles.ActiveLeftIndicatorCharacter
-		ric = m.styles.ActiveRightIndicatorCharacter
-		li = m.styles.ActiveLeftIndicator
-		ri = m.styles.ActiveRightIndicator
-		l = m.styles.ActiveLabel
-	}
-
-	var left, right, label string
-	if lic != "" {
-		left = fmt.Sprintf("%s ", li.Render(lic))
-	}
-	if m.label != "" {
-		label = l.Render(m.label)
-	}
-	if ric != "" {
-		right = fmt.Sprintf(" %s", ri.Render(ric))
-	}
-
-	s := fmt.Sprintf("%s%s%s", left, label, right)
-
-	return s
+	// User-defined data to pass down into the button
+	Data any
 }
 
 // Model represents a group of radio buttons.
@@ -118,14 +34,14 @@ type Model struct {
 	// back to the last button
 	wraparound bool
 
-	// styles for the tab headers and bodies
+	// styles for the radio buttons
 	styles Styles
 }
 
-// New creates a new Model.
+// New creates a new [Model].
 func New(vertical bool, buttons ...tea.Model) Model {
 	// Activate first button
-	b, _ := buttons[0].Update(RadioButtonMsg{true})
+	b, _ := buttons[0].Update(RadioButtonMsg{State: true})
 	buttons[0] = b
 
 	m := Model{
@@ -199,7 +115,7 @@ func (m Model) Init() tea.Cmd { return nil }
 // Update the Model.
 func (m Model) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
 	signalButton := func(i int, state bool) tea.Cmd {
-		n, c := m.buttons[i].Update(RadioButtonMsg{state})
+		n, c := m.buttons[i].Update(RadioButtonMsg{State: state})
 		m.buttons[i] = n
 
 		return c
